+++ tests/test_dc_bridge_p_loaded.rs
@@ -41,10 +41,10 @@ fn build_v() -> Circuit {
     // Diode bridge
     //  (1) is top
     //  (2) is bottom
-    ckt.elements.push( Element::D(Diode{p: 1, n: 3, i_sat: 1e-9, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 4, n: 1, i_sat: 1e-9, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 2, n: 3, i_sat: 1e-9, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 4, n: 2, i_sat: 1e-9, tdegc: 27.0}) );
+    ckt.elements.push( Element::D(Diode::new(1, 3, 1e-9, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(4, 1, 1e-9, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(2, 3, 1e-9, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(4, 2, 1e-9, 27.0)) );
 
     // load
     ckt.elements.push( Element::R(Resistor{a: 3, b: 4, value: 1000.0}) );
@@ -62,10 +62,10 @@ fn build_vv() -> Circuit {
     // Diode bridge
     //  (1) is top
     //  (2) is bottom
-    ckt.elements.push( Element::D(Diode{p: 1, n: 3, i_sat: 1e-9, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 2, n: 1, i_sat: 1e-9, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 0, n: 3, i_sat: 1e-9, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 2, n: 0, i_sat: 1e-9, tdegc: 27.0}) );
+    ckt.elements.push( Element::D(Diode::new(1, 3, 1e-9, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(2, 1, 1e-9, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(0, 3, 1e-9, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(2, 0, 1e-9, 27.0)) );
 
     // load
     ckt.elements.push( Element::R(Resistor{a: 3, b: 2, value: 1000.0}) );