+++ tests/test_trans_ir_bridge_rc_load.rs
@@ -34,10 +34,10 @@ fn build(amp: f32, freq: f32, isat: f32) -> Circuit {
     // Diode bridge
     //  (1) is top
     //  (2) is bottom
-    ckt.elements.push( Element::D(Diode{p: 1, n: 3, i_sat: isat, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 4, n: 1, i_sat: isat, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 2, n: 3, i_sat: isat, tdegc: 27.0}) );
-    ckt.elements.push( Element::D(Diode{p: 4, n: 2, i_sat: isat, tdegc: 27.0}) );
+    ckt.elements.push( Element::D(Diode::new(1, 3, isat, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(4, 1, isat, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(2, 3, isat, 27.0)) );
+    ckt.elements.push( Element::D(Diode::new(4, 2, isat, 27.0)) );
 
     // load
     ckt.elements.push( Element::R(Resistor{a: 3, b: 4, value: 1000.0}) );